[crossorigin contexts] First tests
Just to see if it would be at all doable to also handle cross-origin contexts, and it looks like it is... by using Prox[ies]

Though I'm not quite sure what's the implication of proxyfying Window objects like that. If we can replace all instances, then that should be fine, but we already found out while writing this simple test that a simple `onmessage = ({ source }) => if( source === popup ) { ...` requires to proxify `MessageEvent.source`,  and to use a WeakMap so we always return the SameObejct Proxy for each Window instance we try to proxify.

I guess it boils down to be able to handle all the Window instances, but then, I doubt it's really doable, at least not for `top` which is non-configurable...