Add support for AudioWorklets
AudioWorklets need a special treatment, just like MessageChannel, their MessagePort are not actually created by calling the MessagePort constructor, we can't thus just hook on all MessagePorts `onmessage`; we need to spoof the Interfaces that use these MessagePorts.
We were already doing it for MessageChannel, adding a message handler to both its `.port1` and `.port2` properties, we now are doing it too on AudioWorkletProcessor for its `.port`.

We also removed some debugging remainings that were actually causing an exception in Worklet context, and we don't assume `MessageChannel` to be accessible everywhere, it's not.